requirements.txt
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
vobject==0.9.6.1
